Redirects / to www.joinoneroom.com
We want to keep the static site up and running, but we do not want anyone
visiting the home page of the static site, since what we really want is people
visiting our real home page, which is the course catalog and lives in the app.
Since the app will live at `www.joinoneroom.com` just as soon as we merge these
two, that is where we want this index page to redirect to.